YES! Service and Database connection working w tests
But this is supposed to fail in subsequent runs because of the
database holding state... but this seems to work just fine after
repeated test runs... Hmmm... will have to look into this and see what
the database is holding. But for now going to take a breath and go to
bed.