Fixed: Database Err propagates through service (see: 40846bc5)
Got the system working as of chap3 but it was not failing as it should
when the primary key violation was encountered. It turns out that I
should not have been returning Ok() as an http response during the
error condition, but something more appropriate like BadRequest() or
InternalServerError(). [I missed that]

Only at this point was I able to realize the error scenario described
[here](https://www.lpalmieri.com/posts/2020-08-31-zero-to-production-3-5-html-forms-databases-integration-tests/#:~:text=Wait%2C%20no%2C%20what%20the%20fuck!%20Don%27t%20do%20this%20to%20us!)