Factored out db connection setup in test ...
Following the plan, we want to sequester the set up of the db
connection pool such that we can pass it a modified configuration in
this local scope that will make it create a temporary database and
migrate it. so that each test isolated with clean new state from the
database.

See: (Test Isolation 3.7.1)[https://www.lpalmieri.com/posts/2020-08-31-zero-to-production-3-5-html-forms-databases-integration-tests/#:~:text=3.7.1.%20Test%20Isolation]