Telemetry - now pretty sophisticated... some crazy code...
Okay, this has been a whirlwind tour of telemetry, from straight up
logging (simple, right?) to spans and adjusting the layers' output and
registry.  I still have not wrapped my brain around it.

Note, I am just not adding the telemetry file... but it is assumed to
be present in the previous two commits. I could do some git gymnastics
and make a diff branch and then rebase on the blah blah blah...

* 4eabfbe - (HEAD -> main) Okay this is *weird* - I don't grok this. Seems complicated for no reason, oh well (2 hours ago) <Max G. Faraday>
* 9100ae0 - Telemetry: Refactor code into telemetry module (3 hours ago) <Max G. Faraday>

In 9100ae0 I should have added the initial telemetry file... since
then that file has evolved so it is not even aligned with that commit
to retro add it at this point. Oh well