Moving tracing from within function to instrumentation of the function
This is a pretty cool way to move instrumentation away from the guts
of the function so that you can more cleanly separate concerns.  This
also gives you the method level info, which is what we really wanted
before, but instead of placing it in the function at to the top, we
let tracing::instrument wrap the call.

I like this feature.