Again, moving tracing such that we instrument the db insertion
Further break down the subscribe function so that the query is
sequestered into its own function.  And then instrument (at the
function level) the database interaction.

So at this point the subscribe method is instrumented and simply
delegating to the insert function that does the db interaction.
(Note: no need to use the instrument function off the sqlx::query
function)